project delete confirm dialog

+++ src/app/components/project-list/project-list.component.ts
@@ -1,9 +1,11 @@
+/* tslint:disable:one-line */
 import {Component, OnInit} from '@angular/core';
 import {ProjectService} from '../../core/services/project.service';
 import {Project} from '../../shared/hal-resources/project.resource';
 import {MatDialog, MatSelectChange} from '@angular/material';
 import {ProjectDialogComponent} from '../project-dialog/project-dialog.component';
 import {KeyCloakUser} from '../../keycloak/KeyCloakUser';
+import {MatConfirmDialogComponent} from '../../shared/mat-confirm-dialog/mat-confirm-dialog.component';
 
 @Component({
   selector: 'app-project-list',
@@ -31,11 +33,22 @@ export class ProjectListComponent implements OnInit {
   }
 
   deleteProject(project: Project) {
-    this.projectService.delete(project).subscribe(
-      () => {},
-      error => console.log(error),
-      () => this.getAllProjects()
-    );
+
+    const dialogRef = this.dialog.open(MatConfirmDialogComponent, {
+      data: {title: 'Löschen', message: 'Projekt wirklich löschen?'}
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.projectService.delete(project).subscribe(
+          () => {},
+          error => console.log(error),
+          () => this.getAllProjects()
+        );
+      }
+    });
+
+
   }
 
   getAllProjects() {
